fix: price type change

+++ src/product/Product.java
@@ -7,10 +7,10 @@ public class Product {
     private String name;
     private Category category;
     private String brand;
-    private int price;
+    private long price;
     private int stock;
 
-    public Product(long id, String name, Category category, String brand, int price, int stock) {
+    public Product(long id, String name, Category category, String brand, long price, int stock) {
         this.id = id;
         this.name = name;
         this.category = category;
